+++ play-audio-once.php
@@ -75,3 +75,46 @@ function play_audio_once_assets(): void {
 	}
 }
 add_action( 'enqueue_block_assets', 'play_audio_once_assets' );
+
+/**
+ * Add links in row meta.
+ *
+ * @param array  $links List of links.
+ * @param string $file The requested plugin file name.
+ *
+ * @return array
+ */
+function play_audio_once_add_row_meta_links( array $links, string $file ): array {
+    // bail if this is not our plugin.
+    if ( __FILE__ !== WP_PLUGIN_DIR . DIRECTORY_SEPARATOR . $file ) {
+        return $links;
+    }
+
+    // add our custom links.
+    $row_meta = array(
+        'support' => '<a href="https://wordpress.org/support/plugin/play-audio-once/" target="_blank" title="' . esc_html__( 'Support Forum', 'play-audio-once' ) . '">' . esc_html__( 'Support Forum', 'play-audio-once' ) . '</a>',
+    );
+
+    // return the resulting list of links.
+    return array_merge( $links, $row_meta );
+}
+add_filter( 'plugin_row_meta', 'play_audio_once_add_row_meta_links', 10, 2 );
+
+/**
+ * Add links in plugin list.
+ *
+ * @param array $links List of links on plugin in plugin list.
+ *
+ * @return array
+ */
+function play_audio_once_add_setting_link( array $links ): array {
+    // get language-dependent URL for the how-to.
+    $url = 'https://github.com/threadi/play-audio-once/blob/master/docs/how_to_use.md';
+
+    // add the link to the list.
+    $links[] = '<a href="' . esc_url( $url ) . '" target="_blank">' . esc_html__( 'How to use', 'play-audio-once' ) . '</a>';
+
+    // return resulting list of links.
+    return $links;
+}
+add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), 'play_audio_once_add_setting_link' );